Smart Dashboard/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
jinja2==3.1.2
openai==1.0.0  # Only if you want AI chat
python-dotenv==1.0.0
